+++ ms-notification/internal/service/mocks/mock_service.go
@@ -11,101 +11,10 @@ package mock_service
 
 import (
 	domain "ms-notification/internal/domain"
-	service "ms-notification/internal/service"
 	reflect "reflect"
 
 	gomock "go.uber.org/mock/gomock"
 )
-
-// MockSubscriptionEmails is a mock of SubscriptionEmails interface.
-type MockSubscriptionEmails struct {
-	ctrl     *gomock.Controller
-	recorder *MockSubscriptionEmailsMockRecorder
-	isgomock struct{}
-}
-
-// MockSubscriptionEmailsMockRecorder is the mock recorder for MockSubscriptionEmails.
-type MockSubscriptionEmailsMockRecorder struct {
-	mock *MockSubscriptionEmails
-}
-
-// NewMockSubscriptionEmails creates a new mock instance.
-func NewMockSubscriptionEmails(ctrl *gomock.Controller) *MockSubscriptionEmails {
-	mock := &MockSubscriptionEmails{ctrl: ctrl}
-	mock.recorder = &MockSubscriptionEmailsMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockSubscriptionEmails) EXPECT() *MockSubscriptionEmailsMockRecorder {
-	return m.recorder
-}
-
-// SendConfirmationEmail mocks base method.
-func (m *MockSubscriptionEmails) SendConfirmationEmail(arg0 service.ConfirmationEmailInput) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SendConfirmationEmail", arg0)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// SendConfirmationEmail indicates an expected call of SendConfirmationEmail.
-func (mr *MockSubscriptionEmailsMockRecorder) SendConfirmationEmail(arg0 any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendConfirmationEmail", reflect.TypeOf((*MockSubscriptionEmails)(nil).SendConfirmationEmail), arg0)
-}
-
-// MockWeatherEmails is a mock of WeatherEmails interface.
-type MockWeatherEmails struct {
-	ctrl     *gomock.Controller
-	recorder *MockWeatherEmailsMockRecorder
-	isgomock struct{}
-}
-
-// MockWeatherEmailsMockRecorder is the mock recorder for MockWeatherEmails.
-type MockWeatherEmailsMockRecorder struct {
-	mock *MockWeatherEmails
-}
-
-// NewMockWeatherEmails creates a new mock instance.
-func NewMockWeatherEmails(ctrl *gomock.Controller) *MockWeatherEmails {
-	mock := &MockWeatherEmails{ctrl: ctrl}
-	mock.recorder = &MockWeatherEmailsMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockWeatherEmails) EXPECT() *MockWeatherEmailsMockRecorder {
-	return m.recorder
-}
-
-// SendWeatherForecastDailyEmail mocks base method.
-func (m *MockWeatherEmails) SendWeatherForecastDailyEmail(arg0 service.WeatherForecastEmailInput[*domain.DayWeatherResponse]) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SendWeatherForecastDailyEmail", arg0)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// SendWeatherForecastDailyEmail indicates an expected call of SendWeatherForecastDailyEmail.
-func (mr *MockWeatherEmailsMockRecorder) SendWeatherForecastDailyEmail(arg0 any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendWeatherForecastDailyEmail", reflect.TypeOf((*MockWeatherEmails)(nil).SendWeatherForecastDailyEmail), arg0)
-}
-
-// SendWeatherForecastHourlyEmail mocks base method.
-func (m *MockWeatherEmails) SendWeatherForecastHourlyEmail(arg0 service.WeatherForecastEmailInput[*domain.WeatherResponse]) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SendWeatherForecastHourlyEmail", arg0)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// SendWeatherForecastHourlyEmail indicates an expected call of SendWeatherForecastHourlyEmail.
-func (mr *MockWeatherEmailsMockRecorder) SendWeatherForecastHourlyEmail(arg0 any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendWeatherForecastHourlyEmail", reflect.TypeOf((*MockWeatherEmails)(nil).SendWeatherForecastHourlyEmail), arg0)
-}
 
 // MockEmails is a mock of Emails interface.
 type MockEmails struct {
@@ -132,7 +41,7 @@ func (m *MockEmails) EXPECT() *MockEmailsMockRecorder {
 }
 
 // SendConfirmationEmail mocks base method.
-func (m *MockEmails) SendConfirmationEmail(arg0 service.ConfirmationEmailInput) error {
+func (m *MockEmails) SendConfirmationEmail(arg0 domain.ConfirmationEmailInput) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "SendConfirmationEmail", arg0)
 	ret0, _ := ret[0].(error)
@@ -146,7 +55,7 @@ func (mr *MockEmailsMockRecorder) SendConfirmationEmail(arg0 any) *gomock.Call {
 }
 
 // SendWeatherForecastDailyEmail mocks base method.
-func (m *MockEmails) SendWeatherForecastDailyEmail(arg0 service.WeatherForecastEmailInput[*domain.DayWeatherResponse]) error {
+func (m *MockEmails) SendWeatherForecastDailyEmail(arg0 domain.WeatherForecastEmailInput[*domain.DayWeatherResponse]) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "SendWeatherForecastDailyEmail", arg0)
 	ret0, _ := ret[0].(error)
@@ -160,7 +69,7 @@ func (mr *MockEmailsMockRecorder) SendWeatherForecastDailyEmail(arg0 any) *gomoc
 }
 
 // SendWeatherForecastHourlyEmail mocks base method.
-func (m *MockEmails) SendWeatherForecastHourlyEmail(arg0 service.WeatherForecastEmailInput[*domain.WeatherResponse]) error {
+func (m *MockEmails) SendWeatherForecastHourlyEmail(arg0 domain.WeatherForecastEmailInput[*domain.WeatherResponse]) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "SendWeatherForecastHourlyEmail", arg0)
 	ret0, _ := ret[0].(error)
